add Ics07 Header for UpdateClient Event
