+++ src/module/clients/ics10_grandpa.rs
@@ -1,3 +1,5 @@
+pub mod header;
+
 use ibc::clients::ics10_grandpa::{
 	client_state::ClientState as IbcClientState,
 	help::{
