+++ src/module/clients/mod.rs
@@ -1,1 +1,2 @@
 pub mod ics10_grandpa;
+pub mod ics07_tendermint;
